requirements.txt
ultralytics
opencv-python
numpy
matplotlib
pillow
torch
torchvision
ipykernel
